[kicad] Fix /CART_5V net width
Update the PCB to bring /CART_5V in sync with the schematic.
Apply the power netclass width to /CART_5V as it is currently using a
smaller width than the rest of power nets (from the cartridge pads to the
D2 diode).
Add clearance when needed to account for the increase in net width.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>